+++ src/grimp/application/usecases.py
@@ -1,12 +1,12 @@
 """
 Use cases handle application logic.
 """
-from typing import List
+from typing import List, Set
 
 from ..application.ports.filesystem import AbstractFileSystem
 from ..application.ports.graph import AbstractImportGraph
 from ..application.ports.importscanner import AbstractImportScanner
-from ..application.ports.modulefinder import AbstractModuleFinder
+from ..application.ports.modulefinder import AbstractModuleFinder, FoundPackage
 from ..application.ports.packagefinder import AbstractPackageFinder
 from ..domain.valueobjects import Module
 from .config import settings
@@ -41,30 +41,25 @@ def build_graph(
 
     package_names = [package_name] + list(additional_package_names)
     modules: List[Module] = []
-    modules_by_package_directory = {}
-
+    found_packages: Set[FoundPackage] = set()
     _validate_package_names_are_strings(package_names)
 
     for package_name in package_names:
         package_directory = package_finder.determine_package_directory(
             package_name=package_name, file_system=file_system
         )
-
-        # Build a list of all the Python modules in the package.
-        package_modules = module_finder.find_modules(
+        found_package = module_finder.find_package(
             package_name=package_name,
             package_directory=package_directory,
             file_system=file_system,
         )
-        modules.extend(package_modules)
-        modules_by_package_directory[package_directory] = set(package_modules)
-
-    root_modules = {module.root for module in modules}
+        found_packages.add(found_package)
+        modules.extend(found_package.modules)
 
     import_scanner: AbstractImportScanner = settings.IMPORT_SCANNER_CLASS(
         file_system=file_system,
+        found_packages=found_packages,
         include_external_packages=include_external_packages,
-        modules_by_package_directory=modules_by_package_directory,
     )
     graph: AbstractImportGraph = settings.IMPORT_GRAPH_CLASS()
 
@@ -74,8 +69,10 @@ def build_graph(
         for direct_import in import_scanner.scan_for_imports(module):
             # Before we add the import, check to see if the imported module is in fact an
             # external module, and if so, tell the graph that it is a squashed module.
-            is_external = direct_import.imported.root not in root_modules
-            graph.add_module(direct_import.imported.name, is_squashed=is_external)
+            graph.add_module(
+                direct_import.imported.name,
+                is_squashed=_is_external(direct_import.imported, found_packages),
+            )
 
             graph.add_import(
                 importer=direct_import.importer.name,
@@ -90,4 +87,15 @@ def build_graph(
 def _validate_package_names_are_strings(package_names: List[str]) -> None:
     for name in package_names:
         if not isinstance(name, str):
-            raise TypeError(f"Package names must be strings, got {name.__class__.__name__}.")
+            raise TypeError(
+                f"Package names must be strings, got {name.__class__.__name__}."
+            )
+
+
+def _is_external(module: Module, found_packages: Set[FoundPackage]) -> bool:
+    package_modules = [Module(found_package.name) for found_package in found_packages]
+
+    return not any(
+        module.is_descendant_of(package_module) or module == package_module
+        for package_module in package_modules
+    )