+++ src/grimp/domain/valueobjects.py
@@ -26,13 +26,10 @@ class Module(ValueObject):
             name: The fully qualified name of a Python module, e.g. 'package.foo.bar'.
         """
         self.name = name
+        self.package_name = name.split(".")[0]
 
     def __str__(self) -> str:
         return self.name
-
-    @property
-    def package_name(self) -> str:
-        return self.name.split(".")[0]
 
     @property
     def root(self) -> "Module":